refactor(web): organize file
