+++ apps/web/src/widgets/modal/ui/CreateGroupBoardModal.tsx
@@ -8,6 +8,7 @@ import { Modal, UserItem } from '~/shared/ui';
 import { useModal, useModalState } from '~/shared/hooks';
 import { MODAL, TITLE_MAX_LENGTH } from '~/shared/constants';
 import { TEMPLATE } from '~/widgets/modal/model';
+
 import ImageIcon from '~/assets/icons/image.svg';
 
 export default function CreateGroupBoardModal() {
