+++ apps/web/src/widgets/modal/ui/create-post/CreatePostModal.tsx
