+++ apps/web/src/widgets/modal/ui/create-post/index.ts
@@ -2,3 +2,4 @@ export { default as CreateVote } from './CreateVote';
 export { default as CreatePeerReview } from './CreatePeerReview';
 export { default as CreateMeeting } from './CreateMeeting';
 export { default as CreateNormalPost } from './CreateNormalPost';
+export { default as CreatePostModal } from './CreatePostModal';
