+++ apps/web/src/widgets/modal/ui/create-schedule/CreateScheduleModal.tsx
