+++ apps/web/src/widgets/modal/ui/create-schedule/index.ts
@@ -1,1 +1,2 @@
 export { default as RepeatOption } from './RepeatOption';
+export { default as CreateScheduleModal } from './CreateScheduleModal';
