+++ apps/web/src/widgets/modal/ui/index.ts
@@ -1,15 +1,11 @@
-export { default as CreateProjectModal } from './CreateProjectModal';
-export { default as CreateScheduleModal } from './CreateScheduleModal';
 export { default as LogoutModal } from './LogoutModal';
-export { default as SettingModal } from './SettingModal';
 export { default as AlarmModal } from './AlarmModal';
 export { default as DeleteModal } from './DeleteModal';
 export { default as UpdateModal } from './UpdateModal';
-export { default as CreatePostModal } from './CreatePostModal';
-export { default as ProjectManageModal } from './ProjectManageModal';
-export { default as ProjectMemberManageModal } from './ProjectMemberManageModal';
-export { default as InviteProjectModal } from './InviteProjectModal';
-export { default as ExpulsionProjectModal } from './ExpulsionProjectModal';
-export { default as ChangeProjectModal } from './ChangeProjectModal';
-export { default as LeaveProjectModal } from './LeaveProjectModal';
 export { default as CreateGroupBoardModal } from './CreateGroupBoardModal';
+
+export * from './create-post';
+export * from './create-schedule';
+export * from './project';
+export * from './member';
+export * from './setting';
