+++ apps/web/src/widgets/modal/ui/member/LevelInfoModal.tsx
@@ -5,11 +5,12 @@ import { Button } from '@soup/design-system';
 import { MODAL } from '~/shared/constants';
 import { useModal, useModalState } from '~/shared/hooks';
 import { Modal } from '~/shared/ui';
+
 import {
   LevelTable,
   LevelTableBody,
   LevelTableHead,
-} from './levelTable/LevelTable';
+} from '~/widgets/modal/ui/member';
 
 export default function LevelInfoModal() {
   const { closeModal } = useModal();
