+++ apps/web/src/widgets/modal/ui/project/InviteProjectModal.tsx
