+++ apps/web/src/widgets/modal/ui/project/LeaveProjectModal.tsx
