+++ apps/web/src/widgets/modal/ui/project/ProjectManageModal.tsx
