+++ apps/web/src/widgets/modal/ui/project/index.ts
@@ -1,0 +1,6 @@
+export { default as ChangeProjectModal } from './ChangeProjectModal';
+export { default as CreateProjectModal } from './CreateProjectModal';
+export { default as ExpulsionProjectModal } from './ExpulsionProjectModal';
+export { default as InviteProjectModal } from './InviteProjectModal';
+export { default as LeaveProjectModal } from './LeaveProjectModal';
+export { default as ProjectManageModal } from './ProjectManageModal';
