+++ apps/web/src/widgets/modal/ui/setting/index.ts
@@ -2,3 +2,4 @@ export { default as ProfileSetting } from './ProfileSetting';
 export { default as AlarmSetting } from './AlarmSetting';
 export { default as ToggleController } from './ToggleController';
 export { default as WithdrawSetting } from './WithdrawSetting';
+export { default as SettingModal } from './SettingModal';
